Update store.py
exercise (line 175) is a better variable than just "type", and makes it more obvious as to what it is. line 154 has also been changed from "type" to "exercise", in order to make sure deterioration works properly. measuring (line 176) is either going to be "I" or "M", this allows manipulation of the log later on far easier. all ' have been set to " as well.